+++ Runtime/Audio-Pooling/Scripts/AudioSystem/SoundBuilder.cs
@@ -3,23 +3,33 @@ using UnityEngine;
 
 namespace AudioSystem {
     public class SoundBuilder {
-        readonly SoundManager soundManager;
-        Vector3 position = Vector3.zero;
-        bool randomPitch;
+        readonly SoundManager _soundManager;
+        Vector3 _position = Vector3.zero;
+        private Transform _parent = null;
+        bool _randomPitch;
         private string _cachedName;
         private HashKey _cachedHashKey;
+        private SoundEmitter _cachedEmitter;
+        private SoundData _previousLoopingSoundData;
+        
+        public SoundEmitter GetCachedEmitter => _cachedEmitter;
 
         public SoundBuilder(SoundManager soundManager) {
-            this.soundManager = soundManager;
+            this._soundManager = soundManager;
+        }
+        
+        public SoundBuilder WithParent(Transform parent) {
+            this._parent = parent;
+            return this;
         }
 
         public SoundBuilder WithPosition(Vector3 position) {
-            this.position = position;
+            this._position = position;
             return this;
         }
 
         public SoundBuilder WithRandomPitch() {
-            this.randomPitch = true;
+            this._randomPitch = true;
             return this;
         }
         
@@ -30,14 +40,28 @@ namespace AudioSystem {
             }
             
             if(nameTag == _cachedName) {
-                Play(soundManager.GetSoundData(_cachedHashKey));
+                Play(_soundManager.GetSoundData(_cachedHashKey));
                 return;
             }
 
-            if (!soundManager.TryGetEntryHashString(nameTag, out _cachedHashKey)) return;
+            if (!_soundManager.TryGetEntryHashString(nameTag, out _cachedHashKey)) return;
             
             _cachedName = nameTag;
-            Play(soundManager.GetSoundData(_cachedHashKey));
+            Play(_soundManager.GetSoundData(_cachedHashKey));
+        }
+        
+        public void StopCachedEmitter()
+        {
+            if(!_cachedEmitter) return;
+            if(_cachedEmitter.IsActive) _cachedEmitter.Stop();
+            _previousLoopingSoundData = null;
+            _cachedEmitter = null;
+        }
+        
+        public void SetCachedEmitterVolume(float volume) 
+        {
+            if(!_cachedEmitter) return;
+            if(_cachedEmitter.IsActive) _cachedEmitter.SetVolume(volume);
         }
 
         public void Play(SoundData soundData) {
@@ -46,21 +70,42 @@ namespace AudioSystem {
                 return;
             }
             
-            if (!soundManager.CanPlaySound(soundData)) return;
-            SoundEmitter soundEmitter = soundManager.Get();
+            if (!_soundManager.CanPlaySound(soundData)) return;
+            SoundEmitter soundEmitter = _soundManager.Get();
             soundEmitter.Initialize(soundData);
-            soundEmitter.transform.position = position;
-            soundEmitter.transform.parent = soundManager.transform;
+            soundEmitter.transform.position = _position;
+            soundEmitter.transform.parent = _soundManager.transform;
 
-            if (randomPitch) {
+            if (_randomPitch) {
                 soundEmitter.WithRandomPitch();
             }
 
             if (soundData.frequentSound) {
-                soundEmitter.Node = soundManager.FrequentSoundEmitters.AddLast(soundEmitter);
+                soundEmitter.Node = _soundManager.FrequentSoundEmitters.AddLast(soundEmitter);
             }
             
             soundEmitter.Play();
+        }
+        
+        private void SetCachedEmitter()
+        {
+            if(_cachedEmitter) ResetCachedEmitter();
+            _cachedEmitter = _soundManager.Get();
+            _cachedEmitter.OnStopped += ResetCachedEmitter;
+        }
+
+        private void ResetCachedEmitter()
+        {
+            if(!_cachedEmitter) return;
+            _cachedEmitter.OnStopped -= ResetCachedEmitter;
+            _cachedEmitter = null;
+        }
+
+        private void ClearPreviousLooping()
+        {
+            if (_previousLoopingSoundData == null) return;
+            if(_previousLoopingSoundData.loop && _cachedEmitter)
+                StopCachedEmitter();
         }
     }
 }